Implemented the Quick Action feature from issue  #180
This adds the Quick Convert to GIF with Gifski
quick action in finder, and the settings view which
let's you provide the settings for used when quick
converting.
