+++ Quick Convert to GIF with Gifski/Media.xcassets/TouchBarBezel.colorset/Contents.json
@@ -1,0 +1,14 @@
+{
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  },
+  "colors" : [
+    {
+      "idiom" : "universal",
+      "color" : {
+        "reference" : "systemPurpleColor"
+      }
+    }
+  ]
+}
